moved session logic to home screen
